fix: export function to usageService
