+++ src/app/drive/services/usage.service.ts
@@ -1,6 +1,12 @@
 import { UsageResponse } from '@internxt/sdk/dist/drive/storage/types';
 import { SdkFactory } from '../../core/factory/sdk';
 
+export interface UsageDetailsProps {
+  drive: number;
+  photos: number;
+  backups: number;
+}
+
 export async function fetchUsage(): Promise<UsageResponse> {
   const storageClient = SdkFactory.getInstance().createStorageClient();
   const photosClient = await SdkFactory.getInstance().createPhotosClient();
@@ -15,12 +21,29 @@ export async function fetchUsage(): Promise<UsageResponse> {
   return driveUsage;
 }
 
+async function getUsageDetails(): Promise<UsageDetailsProps> {
+  const storageClient = SdkFactory.getInstance().createStorageClient();
+  const photosClient = await SdkFactory.getInstance().createPhotosClient();
+
+  const [{ drive, backups, total }, { usage: photosUsage }] = await Promise.all([
+    storageClient.spaceUsage(),
+    photosClient.photos.getUsage(),
+  ]);
+
+  return {
+    drive,
+    photos: photosUsage,
+    backups,
+  };
+}
+
 function getUsagePercent(usage: number, limit: number): number {
   return Math.ceil((usage / limit) * 100);
 }
 
 const usageService = {
   fetchUsage,
+  getUsageDetails,
   getUsagePercent,
 };
